@@ -11,7 +11,7 @@
         "groups"
     ],
     "require": {
-        "php": "^7.1",
+        "php": "^7.4",
         "magento/framework": "*",
         "magento/module-catalog": "*",
         "magento/module-checkout": "*",
